Move to xdstest
