+++ pilot/pkg/simulation/traffic.go
@@ -247,7 +247,7 @@ func hasFilterOnPort(l *listener.Listener, filter string, port int) bool {
 	if got.FilterDisabled == nil {
 		return true
 	}
-	return !v1alpha3.EvaluateListenerFilterPredicates(got.FilterDisabled, false, port)
+	return !xdstest.EvaluateListenerFilterPredicates(got.FilterDisabled, false, port)
 }
 
 func (sim *Simulation) Run(input Call) (result Result) {
